apps/python-backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
slowapi>=0.1.9

# AI
anthropic>=0.40.0

# Data / Storage
chromadb-client>=0.5.0
httpx>=0.27.0

# System Monitoring
docker>=7.0.0
psutil>=5.9.0

# Utilities
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
